+++ CHANGELOG.md
@@ -7,7 +7,7 @@ F: fix bug
 * `F` 修复 alert 配置的bug 。
 * `F` calendar 组件，icon 替换成原生的 scss 写的 icon
 * `FA` Captcha 组件，conf example 配置错误。增加未配置的警告。
-* `F` cropper 修复背景 image 。
+* `F` Cropper 修复背景 image 。loading 脱离 spinner 组件。脱离切割和上传完毕之后的配置。
 
 ## 1.3.2 (2017-03-22)